refactor: CustomizedResponseEntityException 클래스 구현

+++ src/main/java/com/sb02/todoapp/controller/TodoRestController.java
@@ -4,9 +4,11 @@ import com.sb02.todoapp.domain.Todo;
 import com.sb02.todoapp.exception.TodoNotFound;
 import com.sb02.todoapp.service.TodoService;
 import jakarta.validation.constraints.NotEmpty;
+import java.net.URI;
 import java.util.List;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.RequestEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -17,6 +19,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 @RestController
 @RequestMapping("/api/todos")
@@ -52,6 +55,13 @@ public class TodoRestController {
 
         todoService.save(todo);
 
+        URI location = ServletUriComponentsBuilder.fromCurrentRequest()
+            .path("/{id}")
+            .buildAndExpand(todo.getId()) // 동적으로 생성된 리소스 아이디를 {id}에 삽입함 .
+            .toUri();
+
+        // 200번대 상태코드 -> 클라이언트 요청을 성공한 경우 .
+        // 상태 코드를 201 create로 바꾸고 생성 위치(location) 반환 .
         return new CreateTodoResponse(todo.getName(), todo.getDescription());
     }
 
@@ -65,12 +75,21 @@ public class TodoRestController {
 
     @DeleteMapping("/{id}")
     public DeleteTodoResponse deleteTodo(@PathVariable("id") Long id){
+
+        Todo todo = todoService.findById(id);
+        if(todo == null){
+            throw new TodoNotFound("todo not found");
+        }
+
         todoService.delete(id);
+
+        // 204 no content 에러 전송 .. -> 삭제 됐으니까 no content
         return new DeleteTodoResponse(id);
     }
 
     @Data
     static class DeleteTodoResponse{
+
         private Long id;
 
         public DeleteTodoResponse(Long id){
